Use a bigger integer to store the result
Kotlin is ... not good at handling unsigned numbers. Therefore u16
becomes i16 and thus we quickly run into problems.
We would still have enougb bits to store it all, but we need conversions
on the Kotlin side, because other values are `Int`.
Plus we already mistakenly used `Int` in Kotlin, where it had to be `Short`.

So let's just switch to `Int` (32-bit) and be done with it.

+++ glean-core/ffi/glean.h
@@ -435,7 +435,7 @@ uint8_t glean_on_ready_to_submit_pings(void);
 
 char *glean_ping_collect(uint64_t ping_type_handle, FfiStr reason);
 
-void glean_process_ping_upload_response(FfiPingUploadTask task, uint16_t status);
+void glean_process_ping_upload_response(FfiPingUploadTask task, uint32_t status);
 
 void glean_quantity_set(uint64_t metric_id, int64_t value);
